requirements.txt
adal==1.2.2
alembic==1.4.1
ansible==2.9.5
appdirs==1.4.3
applicationinsights==0.11.9
argcomplete==1.11.1
arrow==0.15.8
astroid==2.3.3
azure-cli-core==2.0.35
azure-cli-nspkg==3.0.2
azure-common==1.1.11
azure-core==1.8.0
azure-graphrbac==0.40.0
azure-keyvault==1.0.0a1
azure-mgmt-authorization==0.51.1
azure-mgmt-automation==0.1.1
azure-mgmt-batch==5.0.1
azure-mgmt-cdn==3.0.0
azure-mgmt-compute==4.4.0
azure-mgmt-containerinstance==1.4.0
azure-mgmt-containerregistry==2.0.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.5.2
azure-mgmt-devtestlabs==3.0.0
azure-mgmt-dns==2.1.0
azure-mgmt-hdinsight==0.1.0
azure-mgmt-iothub==0.7.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-monitor==0.5.2
azure-mgmt-network==2.3.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-rdbms==1.4.1
azure-mgmt-redis==5.0.0
azure-mgmt-resource==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-sql==0.10.0
azure-mgmt-storage==3.1.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.41.0
azure-nspkg==2.0.0
azure-storage==0.35.1
azure-storage-blob==12.4.0
backcall==0.1.0
bcrypt==3.1.7
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
csv2json==0.2.2
decorator==4.4.2
distlib==0.3.0
filelock==3.0.12
Flask==1.1.1
Flask-API==2.0
Flask-Cors==3.0.8
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.2.0
httplib2==0.17.0
humanfriendly==7.1.1
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.2.0
ipython==7.13.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
knack==0.3.3
kubernetes==11.0.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
msrest==0.6.11
msrestazure==0.5.0
numpy==1.18.2
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.1
paramiko==2.7.1
parso==0.6.2
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==12.7
pytz==2019.3
PyYAML==5.3
qrcode==6.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
simplejson==3.17.2
six==1.14.0
SQLAlchemy==1.3.13
tabulate==0.8.2
telegram==0.0.1
timezonefinder==4.2.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.8
virtualenvwrapper-win==1.2.6
wcwidth==0.1.8
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.0
wrapt==1.11.2
xmltodict==0.12.0
yamllint==1.20.0
zipp==3.1.0
